Refactor activeproject into projectHelpers
